aster_core::fs::file

Trait MappedObject

pub trait MappedObject:
    Send
    + Sync
    + Debug {
    // Required method
    fn dup_at_offset(&self, offset: usize) -> Box<dyn MappedObject>;

    // Provided method
    fn handle_page_fault(
        &self,
        _offset: usize,
        _handle: MapHandle<'_, '_, '_>,
    ) -> Result<()> { ... }
}

A trait that describes memory mapping behavior for special files (after mmap).

Required Methods

fn dup_at_offset(&self, offset: usize) -> Box<dyn MappedObject>

Duplicates the memory mapping at the specific offset.

offset specifies the memory address offset within the mapping, which must be smaller than the mapping size and page-aligned.

Provided Methods

fn handle_page_fault( &self, _offset: usize, _handle: MapHandle<'_, '_, '_>, ) -> Result<()>

Handles the page fault.

offset specifies the memory address offset within the mapping, which must be smaller than the mapping size and page-aligned.

Implementations

impl dyn MappedObject

pub(crate) fn dup(&self) -> Box<dyn MappedObject>

Duplicates the memory mapping.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".