aster_core::fs::fs_impls::devtmpfs

Module tree

Filesystem-tree operations for creating and deleting devtmpfs nodes.

The current devtmpfs backing is a tmpfs-flavored RamFs, so the inodes in this tree are expected to be RamInodes.

Structs

DevtmpfsNode 🔒

The complete description of a device node managed by devtmpfs.

DevtmpfsNodeMeta 🔒

The metadata that describes a devtmpfs node.

InvalidDevtmpfsPath 🔒

An error returned by DevtmpfsNodeMeta::new and DevtmpfsNodeMeta::with_mode.

Functions

create_device_node 🔒

create_node 🔒

delete_node 🔒

lookup_or_create_dir 🔒

lookup_or_create_path 🔒

lookup_path 🔒

matches_device 🔒

remove_empty_parent_dirs 🔒