aster_core::fs::vfs::fs_apis::inode

Trait Inode

pub(crate) trait Inode:
    Any
    + FileOps
    + Send
    + Sync {

    // Required methods
    fn size(&self) -> usize;
    fn resize(&self, self_dentry: &Dentry, new_size: usize) -> Result<()>;
    fn metadata(&self) -> Result<Metadata>;
    fn ino(&self) -> u64;
    fn type_(&self) -> InodeType;
    fn mode(&self) -> Result<InodeMode>;
    fn set_mode(&self, self_dentry: &Dentry, mode: InodeMode) -> Result<()>;
    fn owner(&self) -> Result<Uid>;
    fn set_owner(&self, self_dentry: &Dentry, uid: Uid) -> Result<()>;
    fn group(&self) -> Result<Gid>;
    fn set_group(&self, self_dentry: &Dentry, gid: Gid) -> Result<()>;
    fn atime(&self) -> Duration;
    fn set_atime(&self, self_dentry: &Dentry, time: Duration);
    fn mtime(&self) -> Duration;
    fn set_mtime(&self, self_dentry: &Dentry, time: Duration);
    fn ctime(&self) -> Duration;
    fn set_ctime(&self, self_dentry: &Dentry, time: Duration);
    fn fs(&self) -> Arc<dyn FileSystem> ;
    fn extension(&self) -> &Extension;

    // Provided methods
    fn page_cache(&self) -> Option<Arc<Vmo>> { ... }
    fn create(
        &self,
        self_dentry: &Dentry,
        name: &str,
        type_: InodeType,
        mode: InodeMode,
    ) -> Result<Arc<dyn Inode>> { ... }
    fn create_symlink(
        &self,
        self_dentry: &Dentry,
        name: &str,
        target: &str,
        mode: InodeMode,
    ) -> Result<Arc<dyn Inode>> { ... }
    fn create_tmpfile(
        &self,
        self_dentry: &Dentry,
        mode: InodeMode,
        hard_linkability: HardLinkability,
    ) -> Result<Arc<dyn Inode>> { ... }
    fn mknod(
        &self,
        self_dentry: &Dentry,
        name: &str,
        mode: InodeMode,
        type_: MknodType,
    ) -> Result<Arc<dyn Inode>> { ... }
    fn open(
        &self,
        self_dentry: &Dentry,
        access_mode: AccessMode,
        status_flags: StatusFlags,
    ) -> Option<Result<Box<dyn PerOpenFileOps>>> { ... }
    fn link(
        &self,
        self_dentry: &Dentry,
        old_dentry: &Dentry,
        name: &str,
    ) -> Result<()> { ... }
    fn unlink(&self, child_dentry: &Dentry) -> Result<()> { ... }
    fn rmdir(&self, child_dentry: &Dentry) -> Result<()> { ... }
    fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>> { ... }
    fn rename(
        &self,
        old_child_dentry: &Dentry,
        new_dir_dentry: &Dentry,
        new_name: &str,
        replaced_inode: Option<&Arc<dyn Inode>>,
        mode: RenameMode,
    ) -> Result<()> { ... }
    fn read_link(&self) -> Result<SymbolicLink> { ... }
    fn sync(&self, _mode: SyncMode) -> Result<()> { ... }
    fn fallocate(
        &self,
        mode: FallocMode,
        offset: usize,
        len: usize,
    ) -> Result<()> { ... }
    fn revalidation_policy(&self) -> RevalidationPolicy { ... }
    fn revalidate_exists(&self, _name: &str, _child: &dyn Inode) -> bool { ... }
    fn revalidate_absent(&self, _name: &str) -> bool { ... }
    fn seek_end(&self) -> Option<usize> { ... }
    fn set_xattr(
        &self,
        self_dentry: &Dentry,
        name: XattrName<'_>,
        value_reader: &mut VmReader<'_>,
        flags: XattrSetFlags,
    ) -> Result<()> { ... }
    fn get_xattr(
        &self,
        name: XattrName<'_>,
        value_writer: &mut VmWriter<'_>,
    ) -> Result<usize> { ... }
    fn list_xattr(
        &self,
        namespace: XattrNamespace,
        list_writer: &mut VmWriter<'_>,
    ) -> Result<usize> { ... }
    fn remove_xattr(
        &self,
        self_dentry: &Dentry,
        _name: XattrName<'_>,
    ) -> Result<()> { ... }
    fn check_permission(&self, perm: Permission) -> Result<()> { ... }

}

Represents the filesystem metadata and operations associated with an inode.

When a method accepts self_dentry, it should be the Dentry corresponding to this inode; it provides additional VFS context, such as the parent of the dentry from which the operation originated, which can be useful to some filesystem implementations such as overlayfs.

Required Methods

fn size(&self) -> usize

Returns the current size of the inode.

fn resize(&self, self_dentry: &Dentry, new_size: usize) -> Result<()>

Changes the size of the inode.

fn metadata(&self) -> Result<Metadata>

Returns all inode metadata.

This method returns all metadata fields and is typically used to serve syscalls such as fstat. Callers must not assume that this operation is always cheap. For example, remote filesystems may need to perform I/O when getting or updating metadata, and such operations may fail.

If only a specific field is needed, prefer using the dedicated getters such as size, ino, type_, mode, owner, and group, which are generally cheaper as they return cached data from the inode.

fn ino(&self) -> u64

Returns the inode number.

fn type_(&self) -> InodeType

Returns the inode type.

fn mode(&self) -> Result<InodeMode>

Returns the inode’s permission mode.

fn set_mode(&self, self_dentry: &Dentry, mode: InodeMode) -> Result<()>

Sets the inode’s permission mode.

fn owner(&self) -> Result<Uid>

Returns the inode owner’s user ID.

fn set_owner(&self, self_dentry: &Dentry, uid: Uid) -> Result<()>

Sets the inode owner’s user ID.

fn group(&self) -> Result<Gid>

Returns the inode owner’s group ID.

fn set_group(&self, self_dentry: &Dentry, gid: Gid) -> Result<()>

Sets the inode owner’s group ID.

fn atime(&self) -> Duration

Returns the inode’s access time.

fn set_atime(&self, self_dentry: &Dentry, time: Duration)

Sets the inode’s access time.

fn mtime(&self) -> Duration

Returns the inode’s modification time.

fn set_mtime(&self, self_dentry: &Dentry, time: Duration)

Sets the inode’s modification time.

fn ctime(&self) -> Duration

Returns the inode’s metadata-change time.

fn set_ctime(&self, self_dentry: &Dentry, time: Duration)

Sets the inode’s metadata-change time.

fn fs(&self) -> Arc<dyn FileSystem>

Returns the filesystem containing this inode.

fn extension(&self) -> &Extension

Gets the extension of this inode.

Provided Methods

fn page_cache(&self) -> Option<Arc<Vmo>>

Returns the inode’s page cache, if it has one.

fn create( &self, self_dentry: &Dentry, name: &str, type_: InodeType, mode: InodeMode, ) -> Result<Arc<dyn Inode>>

Creates a child inode.

Use Inode::create_symlink to create symbolic links and Inode::mknod to create device nodes or named pipes.

fn create_symlink( &self, self_dentry: &Dentry, name: &str, target: &str, mode: InodeMode, ) -> Result<Arc<dyn Inode>>

Creates a symbolic-link child inode.

fn create_tmpfile( &self, self_dentry: &Dentry, mode: InodeMode, hard_linkability: HardLinkability, ) -> Result<Arc<dyn Inode>>

Creates an unnamed temporary inode.

fn mknod( &self, self_dentry: &Dentry, name: &str, mode: InodeMode, type_: MknodType, ) -> Result<Arc<dyn Inode>>

Creates a device or named-pipe inode.

fn open( &self, self_dentry: &Dentry, access_mode: AccessMode, status_flags: StatusFlags, ) -> Option<Result<Box<dyn PerOpenFileOps>>>

Opens a file for this inode if supported.

fn link( &self, self_dentry: &Dentry, old_dentry: &Dentry, name: &str, ) -> Result<()>

Creates a hard link to an existing inode.

fn unlink(&self, child_dentry: &Dentry) -> Result<()>

Removes the specified child directory entry from this directory.

fn rmdir(&self, child_dentry: &Dentry) -> Result<()>

Removes the specified child directory from this directory.

fn lookup(&self, name: &str) -> Result<Arc<dyn Inode>>

Looks up a child inode by name.

fn rename( &self, old_child_dentry: &Dentry, new_dir_dentry: &Dentry, new_name: &str, replaced_inode: Option<&Arc<dyn Inode>>, mode: RenameMode, ) -> Result<()>

Renames or moves a child entry, optionally replacing an existing inode.

fn read_link(&self) -> Result<SymbolicLink>

Reads the target of a symbolic link.

fn sync(&self, _mode: SyncMode) -> Result<()>

Flushes inode data and metadata according to the requested synchronization mode.

fn fallocate(&self, mode: FallocMode, offset: usize, len: usize) -> Result<()>

Manipulates a range of space of the file according to the specified allocate mode, the manipulated range starts at offset and continues for len bytes.

fn revalidation_policy(&self) -> RevalidationPolicy

Returns the revalidation policy for cached children of this directory.

See RevalidationPolicy for the full protocol description.

Default: empty (no revalidation). Correct for filesystems where all mutations go through the VFS.

fn revalidate_exists(&self, _name: &str, _child: &dyn Inode) -> bool

Checks whether a cached child inode still exists under this directory.

Called on a positive cache hit when RevalidationPolicy::REVALIDATE_EXISTS is set. Returns true if child is still a valid child named name; false if the entry should be dropped and the lookup retried.

See RevalidationPolicy for the full protocol description.

Precondition

Only called when self.revalidation_policy() includes REVALIDATE_EXISTS. Otherwise, the return value is considered garbage.

fn revalidate_absent(&self, _name: &str) -> bool

Checks whether a name is still absent under this directory.

Called on a negative cache hit when RevalidationPolicy::REVALIDATE_ABSENT is set. Returns true if name is still absent; false if a child may now exist and the lookup should be retried.

See RevalidationPolicy for the full protocol description.

Precondition

Only called when self.revalidation_policy() includes REVALIDATE_ABSENT. Otherwise, the return value is considered garbage.

fn seek_end(&self) -> Option<usize>

Returns the end position for SeekFrom::End.

fn set_xattr( &self, self_dentry: &Dentry, name: XattrName<'_>, value_reader: &mut VmReader<'_>, flags: XattrSetFlags, ) -> Result<()>

Sets an extended attribute on this inode.

fn get_xattr( &self, name: XattrName<'_>, value_writer: &mut VmWriter<'_>, ) -> Result<usize>

Reads an extended attribute value into the supplied writer.

fn list_xattr( &self, namespace: XattrNamespace, list_writer: &mut VmWriter<'_>, ) -> Result<usize>

Lists the extended attributes in the specified namespace.

fn remove_xattr(&self, self_dentry: &Dentry, _name: XattrName<'_>) -> Result<()>

Removes an extended attribute from this inode.

fn check_permission(&self, perm: Permission) -> Result<()>

Used to check for read/write/execute permissions on a file.

Similar to Linux, using “fsuid” here allows setting filesystem permissions without changing the “normal” uids for other tasks.

Implementations

impl dyn Inode

pub(crate) fn downcast_ref<T: Inode>(&self) -> Option<&T>

pub(crate) fn read_bytes_at( &self, offset: usize, buf: &mut [u8], ) -> Result<usize>

pub(crate) fn write_bytes_at(&self, offset: usize, buf: &[u8]) -> Result<usize>