aster_core::fs::vfs::fs_apis

Module xattr

Structs

XattrName 🔒

Represents the name of an xattr. It includes both a valid namespace and a full name string slice, which contains the namespace prefix.

XattrSetFlags 🔒

Flags for setting an xattr value.

Enums

XattrNamespace 🔒

Represents different namespaces with different capabilities for extended attributes (xattrs).

Constants

SECURITY_CAPABILITY_XATTR_NAME 🔒

XATTR_LIST_MAX_LEN 🔒

XATTR_NAME_MAX_LEN 🔒

XATTR_VALUE_MAX_LEN 🔒

Functions

clear_file_priv 🔒

Clears file privileges after an operation modifies file contents.

clear_set_id_bits 🔒