aster_core::process

Module process

Modules

init_proc 🔒

This module defines functions related to spawning the init process.

job_control 🔒

process_group 🔒

session 🔒

terminal 🔒

timer_manager 🔒

Structs

AtomicPid 🔒

Atomic Pid.

ParentProcess 🔒

Representing a parent process by holding a weak reference to it and its PID.

ParentProcessGuard 🔒

Process 🔒

Process stands for a set of threads that shares the same userspace.

ReapedChildrenStats 🔒

The resource usage statistics of child processes that have terminated and been reaped by the parent.

Constants

BOOTSTRAP_PGID 🔒

The process group ID for process group created by ProcessGroup::new_bootstrap.

BOOTSTRAP_SID 🔒

The session ID for the session created by Session::new_bootstrap_pair.

INIT_PROCESS_PID 🔒

The PID of the init process.

Functions

broadcast_signal_async 🔒

Broadcasts a process-directed kernel signal asynchronously.

enqueue_signal_async 🔒

Enqueues a process-directed kernel signal asynchronously.

init_on_each_cpu 🔒

Type Aliases

ExitCode 🔒

Pgid 🔒

Process group ID.

Pid 🔒

Process ID.

Sid 🔒

Session ID.