aster_core

Module syscall

System call handlers.

Modules

accept 🔒

access 🔒

alarm 🔒

arch 🔒

System call dispatch in the x86 architecture.

arch_prctl 🔒

bind 🔒

brk 🔒

capget 🔒

capset 🔒

chdir 🔒

chmod 🔒

chown 🔒

chroot 🔒

clock_gettime 🔒

clone 🔒

close 🔒

connect 🔒

constants 🔒

Constants used in system calls.

dup 🔒

epoll 🔒

eventfd 🔒

eventfd() creates an eventfd object used for event notification.

execve 🔒

exit 🔒

exit_group 🔒

fadvise64 🔒

fallocate 🔒

fcntl 🔒

flock 🔒

fork 🔒

fsconfig 🔒

fsmount 🔒

fsopen 🔒

fsync 🔒

futex 🔒

get_ioprio 🔒

get_priority 🔒

getcpu 🔒

getcwd 🔒

getdents64 🔒

getegid 🔒

geteuid 🔒

getgid 🔒

getgroups 🔒

getpeername 🔒

getpgid 🔒

getpgrp 🔒

getpid 🔒

getppid 🔒

getrandom 🔒

getresgid 🔒

getresuid 🔒

getrusage 🔒

getsid 🔒

getsockname 🔒

getsockopt 🔒

gettid 🔒

gettimeofday 🔒

getuid 🔒

getxattr 🔒

inotify 🔒

ioctl 🔒

kill 🔒

link 🔒

listen 🔒

listmount 🔒

listxattr 🔒

lseek 🔒

madvise 🔒

memfd_create 🔒

mkdir 🔒

mknod 🔒

mmap 🔒

mount 🔒

move_mount 🔒

mprotect 🔒

mremap 🔒

msync 🔒

munmap 🔒

nanosleep 🔒

open 🔒

pause 🔒

personality 🔒

pidfd_getfd 🔒

pidfd_open 🔒

pidfd_send_signal 🔒

pipe 🔒

pivot_root 🔒

poll 🔒

ppoll 🔒

prctl 🔒

pread64 🔒

preadv 🔒

prlimit64 🔒

pselect6 🔒

ptrace 🔒

pwrite64 🔒

pwritev 🔒

read 🔒

readlink 🔒

reboot 🔒

recvfrom 🔒

recvmsg 🔒

removexattr 🔒

rename 🔒

rmdir 🔒

rt_sigaction 🔒

rt_sigpending 🔒

rt_sigprocmask 🔒

rt_sigreturn 🔒

rt_sigsuspend 🔒

rt_sigtimedwait 🔒

sched_affinity 🔒

sched_get_priority_max 🔒

sched_get_priority_min 🔒

sched_getattr 🔒

sched_getparam 🔒

sched_getscheduler 🔒

sched_setattr 🔒

sched_setparam 🔒

sched_setscheduler 🔒

sched_yield 🔒

select 🔒

semctl 🔒

semget 🔒

semop 🔒

sendfile 🔒

sendmmsg 🔒

sendmsg 🔒

sendto 🔒

set_ioprio 🔒

set_priority 🔒

set_robust_list 🔒

set_tid_address 🔒

setdomainname 🔒

setfsgid 🔒

setfsuid 🔒

setgid 🔒

setgroups 🔒

sethostname 🔒

setitimer 🔒

setns 🔒

This module implements the setns syscall.

setpgid 🔒

setregid 🔒

setresgid 🔒

setresuid 🔒

setreuid 🔒

setsid 🔒

setsockopt 🔒

setuid 🔒

setxattr 🔒

shutdown 🔒

sigaltstack 🔒

signalfd 🔒

signalfd implementation for Linux compatibility.

socket 🔒

socketpair 🔒

stat 🔒

statfs 🔒

statx 🔒

symlink 🔒

sync 🔒

sysinfo 🔒

tgkill 🔒

time 🔒

timer_create 🔒

timer_settime 🔒

timerfd_create 🔒

timerfd_gettime 🔒

timerfd_settime 🔒

truncate 🔒

umask 🔒

umount 🔒

uname 🔒

unlink 🔒

unshare 🔒

utimens 🔒

wait4 🔒

waitid 🔒

write 🔒

Macros

dispatch_fn_inner 🔒

impl_syscall_nums_and_dispatch_fn 🔒

log_syscall_entry 🔒

syscall_handler 🔒

This macro is used to define syscall handler. The first param is the number of parameters, The second param is the function name of syscall handler, The third is optional, means the args(if parameter number > 0), The fourth is optional, means if cpu ctx is required.

Structs

SyscallArgument 🔒

Enums

SyscallReturn 🔒

Syscall return

Functions

handle_syscall 🔒